fix: 수수료 단위를 %로 변경 및 DB 필드 추가
- 수수료 입력 필드: (만원) → (%)
- 수수료 타입: INTEGER → DECIMAL(5,2)
- DB 마이그레이션 추가: hired_date, fee, salary 컬럼
- fee 파싱 로직: parseInt → parseFloat

+++ app/api/pipeline/route.ts
@@ -61,7 +61,7 @@ export async function POST(req: NextRequest) {
         next_action: input.next_action || null,
         next_action_date: input.next_action_date || null,
         hired_date: input.hired_date || null,
-        fee: input.fee ? parseInt(input.fee) : null,
+        fee: input.fee ? parseFloat(input.fee) : null,
         salary: input.salary ? parseInt(input.salary) : null,
       })
       .select()
